scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2

supabase>=2.0.0
httpx==0.27.2
httpcore==1.0.4
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
